+++ src/main/java/uk/gov/service/notify/NotificationClientApi.java
@@ -103,6 +103,22 @@ public interface NotificationClientApi {
     LetterResponse sendPrecompiledLetter(String reference, File precompiledPDF) throws NotificationClientException;
 
     /**
+     * The sendPrecompiledLetter method will create an HTTPS POST request. A JWT token will be created and added as an Authorization header to the request.
+     *
+     * @param reference                 A reference specified by the service for the notification. Get all notifications can be filtered by this reference.
+     *                                  This reference can be unique or used used to refer to a batch of notifications.
+     *                                  Cannot be an empty string or null for precompiled PDF files.
+     * @param precompiledPDF            A file containing a PDF conforming to the Notify standards for printing.
+     *                                  The file must be a PDF and cannot be null.
+     * @param postage                   You can choose first or second class postage for your precompiled letter.
+     *                                  Set the value to first for first class, or second for second class. If you do not pass in this argument, the postage will default to second class.
+     * @return <code>LetterResponse</code>
+     *
+     * @throws NotificationClientException see https://docs.notifications.service.gov.uk/java.html#send-a-precompiled-letter-error-codes
+     */
+    LetterResponse sendPrecompiledLetter(String reference, File precompiledPDF, String postage) throws NotificationClientException;
+
+    /**
      * The sendPrecompiledLetterWithInputStream method will create an HTTPS POST request. A JWT token will be created and added as an Authorization header to the request.
      *
      * @param reference                 A reference specified by the service for the notification. Get all notifications can be filtered by this reference.
@@ -115,6 +131,23 @@ public interface NotificationClientApi {
      * @throws NotificationClientException see https://docs.notifications.service.gov.uk/java.html#send-a-precompiled-letter-error-codes
      */
     public LetterResponse sendPrecompiledLetterWithInputStream(String reference, InputStream stream) throws NotificationClientException;
+
+    /**
+     * The sendPrecompiledLetterWithInputStream method will create an HTTPS POST request. A JWT token will be created and added as an Authorization header to the request.
+     *
+     * @param reference                 A reference specified by the service for the notification. Get all notifications can be filtered by this reference.
+     *                                  This reference can be unique or used used to refer to a batch of notifications.
+     *                                  Cannot be an empty string or null for precompiled PDF files.
+     * @param stream                    An <code>InputStream</code> conforming to the Notify standards for printing.
+     *                                  The InputStream cannot be null.
+     * @param postage                   You can choose first or second class postage for your precompiled letter.
+     *                                  Set the value to first for first class, or second for second class. If you do not pass in this argument, the postage will default to second class.
+
+     * @return <code>LetterResponse</code>
+     *
+     * @throws NotificationClientException see https://docs.notifications.service.gov.uk/java.html#send-a-precompiled-letter-error-codes
+     */
+    public LetterResponse sendPrecompiledLetterWithInputStream(String reference, InputStream stream, String postage) throws NotificationClientException;
 
     /**
      * The getNotificationById method will return a <code>Notification</code> for a given notification id.